+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added 
+- New `block` variation
+
 ## [2.61.1] - 2021-09-13
 ### Added
 - New condition in `getNameWithoutVariant` at `pixelHelper.ts` to check if item.name is the same as item.skuName.